requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.28.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
web3>=6.0.0
plotly>=5.17.0

